Address review: extract PollForNoActiveActivities to WorkflowTestBase (#273)
Move the repeated PollForNoActiveActivities helper from EventSubProcessMessageTests
into WorkflowTestBase so future test classes inherit it without duplication.
No behaviour change — same implementation, same signature, same test results.

1/1 EventSubProcessMessage test passing.
